// apps/client/src/app/core/swagger/model/updateProject.ts
/**
 * NgReports
 * NG Reports API
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UpdateProject {
  /**
   * The unique identifier for the project (e.g., a slug).
   */
  projectId?: string;
  /**
   * The name of the project.
   */
  name?: string;
  /**
   * A description of the project.
   */
  description?: string;
}